Move the clipboard handling into the core library
That removes two more functions from the backend interface.
